refactor: ArReviewer to own component
